requirements.txt
requests
keyboard
python-dotenv
